Cache the `Register.get_global_params` classmethod.
This results in a significant performance improvement when registering
many task types.

The cache is cleared whenever a new type is added.  This is no worse than
the previous behavior (where there was no caching at all).  Since the only
way to introduce global parameters is to register a task that has
global parameters, this should be the only situation in which the cache
needs to be cleared.